src/main/java/servlets/formQuestions/CreateFormQuestionsServlet.java
package servlets.formQuestions;

import services.formQuestions.CreateFormQuestionService;

import javax.inject.Inject;
import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

@WebServlet(value = "/api/formQuestions/create")
public class CreateFormQuestionsServlet extends HttpServlet {

    @Inject
    private CreateFormQuestionService service;

    @Override
    public void init() throws ServletException {
        super.init();
    }

    @Override
    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {

        service.execute(req, resp);
    }

}
